+++ CHANGELOG.md
@@ -104,6 +104,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Extended the portfolio holdings to include performance with currency effects for cash positions
+- Added the feature to link an existing token-based user account to a OpenID Connect (`OIDC`) authentication provider
 
 ### Changed